refactor(service): 도메인별 역할 분리를 위한 패키지 구조 변경 (#225)
