+++ backend/demo/src/main/java/com/sesac/solbid/controller/AuthController.java
@@ -5,8 +5,8 @@ import com.sesac.solbid.exception.ErrorCode;
 import com.sesac.solbid.exception.OAuth2Exception;
 import com.sesac.solbid.exception.CustomException;
 import com.sesac.solbid.exception.ReactivationRequiredException;
-import com.sesac.solbid.service.OAuth2Service;
-import com.sesac.solbid.service.PasswordResetService;
+import com.sesac.solbid.service.auth.OAuth2Service;
+import com.sesac.solbid.service.user.PasswordResetService;
 import com.sesac.solbid.util.JwtUtil;
 import com.sesac.solbid.util.CookieUtil;
 import jakarta.servlet.http.HttpServletRequest;
