+++ backend/demo/src/main/java/com/sesac/solbid/controller/UserController.java
@@ -12,8 +12,8 @@ import com.sesac.solbid.dto.user.response.NicknameAvailabilityResponse;
 import com.sesac.solbid.exception.CustomException;
 import com.sesac.solbid.exception.ReactivationRequiredException;
 import com.sesac.solbid.repository.SocialLoginRepository;
-import com.sesac.solbid.service.UserService;
-import com.sesac.solbid.service.SocialUnlinkService;
+import com.sesac.solbid.service.user.UserService;
+import com.sesac.solbid.service.auth.SocialUnlinkService;
 import com.sesac.solbid.util.JwtUtil;
 import com.sesac.solbid.util.CookieUtil;
 import jakarta.servlet.http.Cookie;
