+++ backend/demo/src/main/java/com/sesac/solbid/service/auth/OAuth2UrlGenerator.java
@@ -1,4 +1,4 @@
-package com.sesac.solbid.service;
+package com.sesac.solbid.service.auth;
 
 import com.sesac.solbid.exception.OAuth2Exception;
 import com.sesac.solbid.exception.ErrorCode;
