+++ backend/demo/src/main/java/com/sesac/solbid/service/user/InMemoryPasswordResetTokenService.java
@@ -1,4 +1,4 @@
-package com.sesac.solbid.service;
+package com.sesac.solbid.service.user;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Profile;
