+++ backend/demo/src/test/java/com/sesac/solbid/controller/AuthControllerTest.java
@@ -7,8 +7,8 @@ import com.sesac.solbid.dto.user.response.LoginResponse;
 import com.sesac.solbid.exception.ErrorCode;
 import com.sesac.solbid.exception.GlobalExceptionHandler;
 import com.sesac.solbid.exception.OAuth2Exception;
-import com.sesac.solbid.service.OAuth2Service;
-import com.sesac.solbid.service.PasswordResetService;
+import com.sesac.solbid.service.auth.OAuth2Service;
+import com.sesac.solbid.service.user.PasswordResetService;
 import com.sesac.solbid.dto.auth.response.AuthUrlResponse;
 import com.sesac.solbid.dto.auth.request.CallbackRequest;
 import com.sesac.solbid.util.CookieUtil;
