+++ backend/demo/src/test/java/com/sesac/solbid/service/OAuth2ServiceIntegrationTest.java
@@ -3,6 +3,10 @@ package com.sesac.solbid.service;
 import com.sesac.solbid.domain.User;
 import com.sesac.solbid.exception.ErrorCode;
 import com.sesac.solbid.exception.OAuth2Exception;
+import com.sesac.solbid.service.auth.OAuth2Service;
+import com.sesac.solbid.service.auth.OAuth2StateService;
+import com.sesac.solbid.service.auth.OAuth2UrlGenerator;
+import com.sesac.solbid.service.user.UserService;
 import com.sesac.solbid.util.JwtUtil;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
