+++ backend/demo/src/test/java/com/sesac/solbid/service/UserServiceSocialLoginTest.java
@@ -3,10 +3,9 @@ package com.sesac.solbid.service;
 import com.sesac.solbid.domain.SocialLogin;
 import com.sesac.solbid.domain.User;
 import com.sesac.solbid.domain.enums.ProviderType;
-import com.sesac.solbid.domain.enums.UserStatus;
-import com.sesac.solbid.domain.enums.UserType;
 import com.sesac.solbid.repository.SocialLoginRepository;
 import com.sesac.solbid.repository.UserRepository;
+import com.sesac.solbid.service.user.UserService;
 import com.sesac.solbid.util.JwtUtil;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
